add get order

+++ internal/model/order_detail.go
@@ -4,7 +4,7 @@ type OrderDetail struct {
 	ID        int64   `db:"id"`
 	OrderID   int64   `db:"order_id"`
 	ProductID int64   `db:"product_id"`
-	Quantity  int     `db:"quantity"`
+	Quantity  int64   `db:"quantity"`
 	Price     float64 `db:"price"`
 	SubTotal  float64 `db:"sub_total"`
 }
